feat(ndh): add a new version for niaid
