+++ src/Homepage/AmbiHomepage.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import { RelayProjectDashboard } from './RelayHomepage';
-import ReduxProjectDashboard from './ReduxProjectDashboard';
+import { RelayProjectDashboard } from './Relay/RelayHomepage';
+import ReduxProjectDashboard from './Redux/ReduxProjectDashboard';
+import { RelayTransactionLogComponent } from "./RelayTransactionLogComponent";
 import { getReduxStore } from '../reduxStore';
 
 
@@ -31,10 +32,20 @@ class AmbidextrousDashboard extends React.Component {
     const nowMs = Date.now();
     // use redux if it has the data, and data is fresh
     if (dataInReduxDate && (nowMs - dataInReduxDate.getTime() < 300000)) {
-      return <ReduxProjectDashboard />;
+      return (
+        <div>
+          <ReduxProjectDashboard />
+          <RelayTransactionLogComponent />
+        </div>
+      );
     }
     dataInReduxDate = new Date();
-    return <RelayProjectDashboard />;
+    return (
+      <div>
+        <RelayProjectDashboard />
+        <RelayTransactionLogComponent />
+      </div>
+    );
   }
 }
